better regex for image
